+++ schug/endpoints/exons.py
@@ -2,12 +2,13 @@ from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 from fastapi.responses import StreamingResponse
+from sqlmodel import Session, select
+
 from schug.database.session import get_session
-from schug.load.ensembl import fetch_ensembl_exons
+from schug.load.ensembl import CHROMOSOMES, fetch_ensembl_exons
 from schug.load.fetch_resource import stream_resource
 from schug.models import Exon, ExonRead
 from schug.models.common import Build
-from sqlmodel import Session, select
 
 router = APIRouter()
 
@@ -27,7 +28,17 @@ def read_exons(
 async def ensembl_exons(build: Build):
     """A proxy to the Ensembl Biomart that retrieves exons in a specific genome build."""
 
-    ensembl_client: EnsemblBiomartClient = fetch_ensembl_exons(build=build)
-    url: str = ensembl_client.build_url(xml=ensembl_client.xml)
+    async def chromosome_stream():
+        for chrom in CHROMOSOMES:
+            print(f"Retrieving exons from chromosome: {chrom}")
+            ensembl_client: EnsemblBiomartClient = fetch_ensembl_exons(
+                build=build, chromosomes=[chrom]
+            )
+            url: str = ensembl_client.build_url(xml=ensembl_client.xml)
 
-    return StreamingResponse(stream_resource(url), media_type="text/tsv")
+            # Stream each chunk from the resource
+            async for chunk in stream_resource(url):
+                yield chunk
+
+    # Return the StreamingResponse with the asynchronous generator
+    return StreamingResponse(chromosome_stream(), media_type="text/tsv")