+++ schug/endpoints/transcripts.py
@@ -2,14 +2,15 @@ from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 from fastapi.responses import StreamingResponse
+from sqlmodel import Session, select
+
 from schug.database.session import get_session
 from schug.load.biomart import EnsemblBiomartClient
-from schug.load.ensembl import fetch_ensembl_transcripts
+from schug.load.ensembl import CHROMOSOMES, fetch_ensembl_transcripts
 from schug.load.fetch_resource import stream_resource
 from schug.models import Transcript, TranscriptRead
 from schug.models.common import Build
 from schug.models.transcript import TranscriptReadWithExons
-from sqlmodel import Session, select
 
 router = APIRouter()
 
@@ -43,7 +44,17 @@ def read_transcript_db_id(
 async def ensembl_transcripts(build: Build):
     """A proxy to the Ensembl Biomart that retrieves transcripts in a specific genome build."""
 
-    ensembl_client: EnsemblBiomartClient = fetch_ensembl_transcripts(build=build)
-    url: str = ensembl_client.build_url(xml=ensembl_client.xml)
+    async def chromosome_stream():
+        for chrom in CHROMOSOMES:
+            print(f"Retrieving transcripts from chromosome: {chrom}")
+            ensembl_client: EnsemblBiomartClient = fetch_ensembl_transcripts(
+                build=build, chromosomes=[chrom]
+            )
+            url: str = ensembl_client.build_url(xml=ensembl_client.xml)
 
-    return StreamingResponse(stream_resource(url), media_type="text/tsv")
+            # Stream each chunk from the resource
+            async for chunk in stream_resource(url):
+                yield chunk
+
+    # Return the StreamingResponse with the asynchronous generator
+    return StreamingResponse(chromosome_stream(), media_type="text/tsv")